chore: change camle case files to kabab case
