+++ components/shared/get-started/fields/availability/availability-header.tsx
@@ -4,7 +4,7 @@ import { LuPlus } from "react-icons/lu";
 import {
 
   DialogTrigger,
-} from "./TimeDialog";
+} from "./time-dialog";
 import { AiTwotoneCheckCircle } from "react-icons/ai";
 import { AvailabilityHeaderTypes } from "@/types/get-started";
