+++ components/shared/get-started/fields/availability/day-fields.tsx
