+++ components/shared/get-started/fields/availability/hour-field.tsx
@@ -1,42 +1,32 @@
-import { Button } from "@/components/ui/button";
+import { Button } from '@/components/ui/button';
 import {
   Command,
   CommandEmpty,
   CommandGroup,
-  CommandInput,
   CommandItem,
   CommandList,
-} from "./TimeCommand";
+} from "./time-command";
 import {
-  Form,
   FormControl,
-  FormDescription,
   FormField,
   FormItem,
-  FormLabel,
   FormMessage,
 } from "@/components/ui/form";
-import { Popover, PopoverContent, PopoverTrigger } from "./TimePopover";
+import { Popover, PopoverContent, PopoverTrigger } from "./time-popover";
 import { cn } from "@/lib/utils";
 import { Check, ChevronsUpDown } from "lucide-react";
 import { getStartedForm } from "@/data/get-started-data";
-import { DayPeriodFieldProps } from "@/types/get-started";
+import { HourFieldProps } from "@/types/get-started";
 
-export function DayPeriodField({
-  control,
-  type,
-  form,
-  day,
-  timeError
-}: DayPeriodFieldProps) {
+export function HourField({ control, type, form, day, timeError }: HourFieldProps) {
   return (
     <FormField
       control={control}
       name={`availability.${day.value}${
-        type === "start" ? ".startTime.dayPeriod" : ".endTime.dayPeriod"
-      }`}
+        type === "start" ? ".startTime.hour" : ".endTime.hour"
+      }`} // sets the name to the specific form data value
       render={({ field }) => (
-        <FormItem className="flex flex-col ml-[3px]">
+        <FormItem className="flex flex-col">
           <Popover>
             <PopoverTrigger asChild>
               <FormControl>
@@ -44,8 +34,8 @@ export function DayPeriodField({
                   variant="outline"
                   role="combobox"
                   className={cn(
-                    " w-16 px-2 gap-0",
-                    !field.value && "text-muted-foreground"
+                    ' w-16 px-2 gap-0',
+                    !field.value && 'text-muted-foreground'
                   )}
                 >
                   {field.value}
@@ -56,32 +46,32 @@ export function DayPeriodField({
             <PopoverContent className="w-16 p-0">
               <Command>
                 <CommandList>
-                  <CommandEmpty>No dayPeriod found.</CommandEmpty>
+                  <CommandEmpty>No hours found.</CommandEmpty>
                   <CommandGroup>
-                    {getStartedForm.dayPeriods.map((dayPeriod) => (
+                    {getStartedForm.hours.map((hour) => (
                       <CommandItem
-                        value={dayPeriod.label}
-                        key={dayPeriod.value}
+                        value={hour.label}
+                        key={hour.value}
                         onSelect={() => {
                           form.setValue(
                             `availability.${day.value}${
                               type === "start"
-                                ? ".startTime.dayPeriod"
-                                : ".endTime.dayPeriod"
+                                ? ".startTime.hour"
+                                : ".endTime.hour"
                             }`,
-                            dayPeriod.value
-                          ); // sets the name to the specific form data value
+                            hour.value
+                          ); // sets the hour value selected on the popout on the form
                           timeError()
                         }}
                         className=" cursor-pointer"
                       >
-                        {dayPeriod.label}
+                        {hour.label}
                         <Check
                           className={cn(
-                            "ml-[-5px]",
-                            dayPeriod.value === field.value
-                              ? "opacity-100"
-                              : "opacity-0"
+                            'ml-[-5px]',
+                            hour.value === field.value
+                              ? 'opacity-100'
+                              : 'opacity-0'
                           )}
                         />
                       </CommandItem>
