+++ components/shared/get-started/fields/availability/minute-field.tsx
@@ -6,14 +6,14 @@ import {
   CommandItem,
   CommandList,
 
-} from "./TimeCommand";
+} from "./time-command";
 import {
   FormControl,
   FormField,
   FormItem,
   FormMessage,
 } from "@/components/ui/form";
-import { Popover, PopoverContent, PopoverTrigger } from "./TimePopover";
+import { Popover, PopoverContent, PopoverTrigger } from "./time-popover";
 import { cn } from "@/lib/utils";
 import { Check, ChevronsUpDown } from "lucide-react";
 import { getStartedForm } from "@/data/get-started-data";
