+++ components/shared/get-started/main/GetStartedCard.tsx
