+++ components/shared/get-started/main/GetStartedSidebar.tsx
