+++ components/views/get-started.tsx
@@ -1,6 +1,6 @@
 import { GetStartedProps } from '@/types/get-started';
-import GetStartedCard from '../shared/GetStarted/main/GetStartedCard';
-import GetStartedSidebar from '../shared/GetStarted/main/GetStartedSidebar';
+import GetStartedCard from '../shared/get-started/main/GetStartedCard';
+import GetStartedSidebar from '../shared/get-started/main/GetStartedSidebar';
 
 const GetStartedView = ({ type }: GetStartedProps) => {
   return (
